refactor: move one test case
